Split generate_api test modules
Move nested generate_api test modules into sibling tests.rs files so
implementation files stay smaller and easier to work in without losing
coverage.

Tighten AGENTS.md guidance so the local instructions stay concise,
reviewable, and explicit about keeping tests separate when practical.

Copilot-Session: 5e82b158-26b2-4ba6-bf85-f1a08ab3aa5c
